revert: move utils from shared back to api
