+++ packages/api/src/activate/utils.ts
