+++ packages/api/src/auth/utils.ts
@@ -1,6 +1,17 @@
 import { UnauthorizedError } from '@error/errors';
 import { Request } from 'express';
 
+export const generateLoginMessage = (
+  account: string,
+  nonce: string
+): string => {
+  return (
+    'SIGN THIS MESSAGE TO LOGIN TO PRAISE.\n\n' +
+    `ADDRESS:\n${account}\n\n` +
+    `NONCE:\n${nonce}`
+  );
+};
+
 export const extractAccessTokenFromRequest = (req: Request): string => {
   const AuthHeader = req.headers['authorization'];
   if (typeof AuthHeader === 'undefined')
