--- packages/shared/src/auth/utils.ts
@@ -1,10 +1,0 @@
-export const generateLoginMessage = (
-  account: string,
-  nonce: string
-): string => {
-  return (
-    'SIGN THIS MESSAGE TO LOGIN TO PRAISE.\n\n' +
-    `ADDRESS:\n${account}\n\n` +
-    `NONCE:\n${nonce}`
-  );
-};
